+++ src/app/(main)/writeDiary/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 // * import component
 import { Button, TextField, Typography, Box } from '@mui/material'
-import DiaryWritePageUI from '@/components/DiaryWritePageUI'
+import DiaryWritePageUI from '@/app/(main)/writeDiary/_component/DiaryWritePageUI'
 
 // * import from next
 // import { useRouter } from 'next/router'
