+++ src/app/layout.tsx
@@ -6,15 +6,15 @@ import '@fontsource/roboto/700.css'
 import CssBaseline from '@mui/material/CssBaseline'
 
 // * import contextProvider
-import { AuthContextProvider } from '@/components/AuthContextProvider'
+import { AuthContextProvider } from '@/shared/context/AuthContextProvider'
 
 // const inter = Inter({ subsets: ['latin'] })
 import Container from '@mui/material/Container'
 import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter'
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Simple Diary',
+  description: 'Write your diary easily By Simple Diary',
 }
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
