+++ src/shared/components/LoginButton.tsx
@@ -4,7 +4,7 @@
 import { Button } from '@mui/material'
 
 // * import api
-import { login, addFirstVisitUser } from '@/api/api'
+import { login, addFirstVisitUser } from '@/shared/api/api'
 
 // * import routing
 import { useRouter } from 'next/navigation'
@@ -13,16 +13,16 @@ import { useRouter } from 'next/navigation'
 import { useContext } from 'react'
 
 // * import type
-import { AuthContextValue } from '@/type/type'
+import { AuthContextValue } from '@/shared/types/type'
 
 // * import context
-import { AuthContext } from '@/context/AuthContext'
+import { AuthContext } from '@/shared/context/AuthContext'
 
 // * import from next
 import Image from 'next/image'
 
 // * import utils
-import { Utils } from '@/utils/utility'
+import { Utils } from '@/shared/utils/utility'
 
 export const LoginButton = () => {
   const context = useContext(AuthContext)
@@ -30,6 +30,8 @@ export const LoginButton = () => {
   return (
     <Button
       onClick={async () => {
+        console.log('1')
+
         login().then((res: AuthContextValue | null) => {
           if (res) {
             console.log(`${res.displayName}님 환영합니다😊`)
@@ -39,7 +41,7 @@ export const LoginButton = () => {
               validDays: 100,
             })
             addFirstVisitUser(res)
-            router.push('/main')
+            router.push('/')
           }
         })
       }}
