+++ src/shared/context/AuthContext.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 // * import type
-import { IAuthContext } from '@/type/type'
+import { IAuthContext } from '@/shared/types/type'
 
 // * import context관련
 import { createContext } from 'react'
